+++ doc/conf.py
@@ -26,6 +26,11 @@ import cadquery
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 # sys.path.insert(0, os.path.abspath('.'))
 
+
+def setup(app):
+    app.add_css_file("tables.css")
+
+
 # -- General configuration -----------------------------------------------------
 
 # If your documentation needs a minimal Sphinx version, state it here.